Revert "[WIP] Terraform Rollback (#3525)" (#3676)
* Revert "[WIP] Terraform Rollback (#3525)"

This reverts commit 26dda08c8656f79080d058c99ff483a50aa0a228.

* Fix: serialization_classes

+++ 71-rest/src/main/java/software/wings/delegatetasks/TerraformProvisionTask.java
@@ -27,7 +27,6 @@ import software.wings.beans.GitConfig;
 import software.wings.beans.GitConfig.GitRepositoryType;
 import software.wings.beans.command.CommandExecutionResult.CommandExecutionStatus;
 import software.wings.beans.delegation.TerraformProvisionParameters;
-import software.wings.beans.delegation.TerraformProvisionParameters.TerraformCommandUnit;
 import software.wings.security.encryption.EncryptedDataDetail;
 import software.wings.service.impl.yaml.GitClientHelper;
 import software.wings.service.intfc.FileService.FileBucket;
@@ -56,6 +55,7 @@ public class TerraformProvisionTask extends AbstractDelegateRunnableTask {
   private static final Logger logger = LoggerFactory.getLogger(TerraformProvisionTask.class);
   private static final String TERRAFORM_STATE_FILE_NAME = "terraform.tfstate";
   private static final String TERRAFORM_VARIABLES_FILE_NAME = "terraform.tfvars";
+  private static final String TERRAFORM_OUTPUTS_FILE_NAME = "terraform.tfouts";
 
   @Inject private GitClient gitClient;
   @Inject private GitClientHelper gitClientHelper;
@@ -126,7 +126,7 @@ public class TerraformProvisionTask extends AbstractDelegateRunnableTask {
 
           if (isNotEmpty(parameters.getEncryptedVariables())) {
             for (Entry<String, EncryptedDataDetail> entry : parameters.getEncryptedVariables().entrySet()) {
-              String value = String.valueOf(encryptionService.getDecryptedValue(entry.getValue()));
+              String value = new String(encryptionService.getDecryptedValue(entry.getValue()));
               saveVariable(writer, entry.getKey(), value);
             }
           }
@@ -138,18 +138,13 @@ public class TerraformProvisionTask extends AbstractDelegateRunnableTask {
       File tfOutputsFile = Paths.get(scriptDirectory, TERRAFORM_VARIABLES_FILE_NAME).toFile();
 
       String joinedCommands = null;
-      switch (parameters.getCommand()) {
-        case APPLY:
-          joinedCommands = Joiner.on(" && ").join(asList("cd " + scriptDirectory, "terraform init -input=false",
-              "terraform refresh -input=false", "terraform plan -out=tfplan -input=false",
-              "terraform apply -input=false tfplan", "(terraform output --json > " + tfOutputsFile.toString() + ")"));
-          break;
-        case DESTROY:
-          joinedCommands = Joiner.on(" && ").join(asList("cd " + scriptDirectory, "terraform init -input=false",
-              "terraform refresh -input=false", "terraform destroy -force"));
-          break;
-        default:
-          throw new IllegalArgumentException("Invalid Terraform Command : " + parameters.getCommand().name());
+      if ("Apply".equals(parameters.getCommandUnitName())) {
+        joinedCommands = Joiner.on(" && ").join(asList("cd " + scriptDirectory, "terraform init -input=false",
+            "terraform refresh -input=false", "terraform plan -out=tfplan -input=false",
+            "terraform apply -input=false tfplan", "(terraform output --json > " + tfOutputsFile.toString() + ")"));
+      } else if ("Destroy".equals(parameters.getCommandUnitName())) {
+        joinedCommands = Joiner.on(" && ").join(asList("cd " + scriptDirectory, "terraform init -input=false",
+            "terraform refresh -input=false", "terraform destroy -force"));
       }
 
       List<Boolean> instances = asList(false);
@@ -208,11 +203,10 @@ public class TerraformProvisionTask extends AbstractDelegateRunnableTask {
           TerraformExecutionData.builder()
               .entityId(delegateFile.getEntityId())
               .stateFileId(delegateFile.getFileId())
-              .terraformProvisionParameters(parameters)
               .executionStatus(code == 0 ? ExecutionStatus.SUCCESS : ExecutionStatus.FAILED)
               .errorMessage(code == 0 ? null : "The terraform command exited with code " + code);
 
-      if (parameters.getCommandUnit() != TerraformCommandUnit.Destroy) {
+      if (!"Destroy".equals(parameters.getCommandUnitName())) {
         terraformExecutionDataBuilder.outputs(new String(Files.readAllBytes(tfOutputsFile.toPath()), Charsets.UTF_8));
       }
 
@@ -243,7 +237,7 @@ public class TerraformProvisionTask extends AbstractDelegateRunnableTask {
             .withAppId(parameters.getAppId())
             .withActivityId(parameters.getActivityId())
             .withLogLevel(INFO)
-            .withCommandUnitName(parameters.getCommandUnit().name())
+            .withCommandUnitName(parameters.getCommandUnitName())
             .withLogLine(line)
             .withExecutionResult(commandExecutionStatus)
             .build());
